chore: remove unused tests and fixtures
