--- fixtures/configure/playground/superface/swapidev.provider.corupted.json
@@ -1,18 +1,0 @@
-{
-  "name": "test",
-  "services": [
-    {
-      "baseUrl": "https://swapi.dev/api",
-      "id": "test"
-    }
-  ],
-  "securitySchemes": [
-    {
-      "id": "test"
-      "type": "apiKey",
-      "in": "header",
-      "name": "X-API-Key"
-    }
-  ],
-  "defaultService": "test"
-}
