--- fixtures/consumer/starwars/.gitignore
@@ -1,4 +1,0 @@
-package-lock.json
-.npmrc
-.yalc
-node_modules
