--- fixtures/consumer/starwars/package.json
@@ -1,11 +1,0 @@
-{
-  "name": "playground",
-  "private": true,
-  "dependencies": {
-    "@superfaceai/one-sdk": "0.0.17"
-  },
-  "devDependencies": {
-    "@types/node": "^14",
-    "typescript": "^4"
-  }
-}
