--- fixtures/consumer/starwars/superface/.gitignore
@@ -1,4 +1,0 @@
-build
-node_modules
-package-lock.json
-
