--- fixtures/consumer/starwars/superface/grid/starwars/character-information.swapidev-ela@1.0.suma
@@ -1,32 +1,0 @@
-profile = "starwars/character-information@1.0"
-provider = "swapidev-ela"
-
-map RetrieveCharacterInformation {
-  http GET "/people/" {
-    request {
-      query {
-        search = input.characterName
-      }
-    }
-
-    response 200 "application/json" {
-      return map error if (body.count === 0) {
-        message = "no character found"
-
-      }
-
-      entries = body.results.filter((result) => result.name === input.characterName)
-      map error if (entries === undefined || entries.length === 0) {
-        message = "not found"
-      }
-      
-      character = entries[0]
-      map result {
-        name = character.name
-        height = character.height
-        weight = character.mass
-        yearOfBirth = character.birth_year
-      }
-    }
-  }
-}
