--- fixtures/consumer/starwars/superface/grid/starwars/character-information@1.0.0.supr
@@ -1,18 +1,0 @@
-name = "starwars/character-information"
-version = "1.0.0"
-
-usecase RetrieveCharacterInformation safe {
-  input {
-    characterName
-  }
-
-  result {
-    height
-    weight
-    yearOfBirth
-  }
-
-  error {
-    message
-  }
-}
