--- fixtures/consumer/starwars/superface/grid/starwars/character-information@1.0.0.supr.ast.json
@@ -1,181 +1,0 @@
-{
-    "definitions": [
-        {
-            "error": {
-                "kind": "UseCaseSlotDefinition",
-                "location": {
-                    "column": 3,
-                    "line": 22
-                },
-                "span": {
-                    "end": 224,
-                    "start": 199
-                },
-                "type": {
-                    "fields": [
-                        {
-                            "fieldName": "message",
-                            "kind": "FieldDefinition",
-                            "location": {
-                                "column": 5,
-                                "line": 24
-                            },
-                            "required": false,
-                            "span": {
-                                "end": 219,
-                                "start": 212
-                            }
-                        }
-                    ],
-                    "kind": "ObjectDefinition",
-                    "location": {
-                        "column": 9,
-                        "line": 22
-                    },
-                    "span": {
-                        "end": 224,
-                        "start": 205
-                    }
-                }
-            },
-            "input": {
-                "kind": "UseCaseSlotDefinition",
-                "location": {
-                    "column": 3,
-                    "line": 6
-                },
-                "span": {
-                    "end": 137,
-                    "start": 106
-                },
-                "type": {
-                    "fields": [
-                        {
-                            "fieldName": "characterName",
-                            "kind": "FieldDefinition",
-                            "location": {
-                                "column": 5,
-                                "line": 8
-                            },
-                            "required": false,
-                            "span": {
-                                "end": 132,
-                                "start": 119
-                            }
-                        }
-                    ],
-                    "kind": "ObjectDefinition",
-                    "location": {
-                        "column": 9,
-                        "line": 6
-                    },
-                    "span": {
-                        "end": 137,
-                        "start": 112
-                    }
-                }
-            },
-            "kind": "UseCaseDefinition",
-            "location": {
-                "column": 1,
-                "line": 5
-            },
-            "result": {
-                "kind": "UseCaseSlotDefinition",
-                "location": {
-                    "column": 3,
-                    "line": 12
-                },
-                "span": {
-                    "end": 195,
-                    "start": 141
-                },
-                "type": {
-                    "fields": [
-                        {
-                            "fieldName": "height",
-                            "kind": "FieldDefinition",
-                            "location": {
-                                "column": 5,
-                                "line": 14
-                            },
-                            "required": false,
-                            "span": {
-                                "end": 161,
-                                "start": 155
-                            }
-                        },
-                        {
-                            "fieldName": "weight",
-                            "kind": "FieldDefinition",
-                            "location": {
-                                "column": 5,
-                                "line": 16
-                            },
-                            "required": false,
-                            "span": {
-                                "end": 173,
-                                "start": 167
-                            }
-                        },
-                        {
-                            "fieldName": "yearOfBirth",
-                            "kind": "FieldDefinition",
-                            "location": {
-                                "column": 5,
-                                "line": 18
-                            },
-                            "required": false,
-                            "span": {
-                                "end": 190,
-                                "start": 179
-                            }
-                        }
-                    ],
-                    "kind": "ObjectDefinition",
-                    "location": {
-                        "column": 10,
-                        "line": 12
-                    },
-                    "span": {
-                        "end": 195,
-                        "start": 148
-                    }
-                }
-            },
-            "safety": "safe",
-            "span": {
-                "end": 226,
-                "start": 60
-            },
-            "useCaseName": "RetrieveCharacterInformation"
-        }
-    ],
-    "header": {
-        "kind": "ProfileHeader",
-        "location": {
-            "column": 1,
-            "line": 1
-        },
-        "name": "character-information",
-        "scope": "starwars",
-        "span": {
-            "end": 57,
-            "start": 0
-        },
-        "version": {
-            "major": 1,
-            "minor": 0,
-            "patch": 0
-        }
-    },
-    "kind": "ProfileDocument",
-    "location": {
-        "column": 1,
-        "line": 1
-    },
-    "span": {
-        "end": 226,
-        "start": 0
-    }
-}
