--- fixtures/consumer/starwars/superface/super.json
@@ -1,17 +1,0 @@
-{
-  "profiles": {
-    "starwars/character-information": {
-      "file": "grid/starwars/character-information@1.0.0.supr",
-      "version": "1.0.0",
-      "providers": {
-        "swapidev-ela": {
-        }
-      }
-    }
-  },
-  "providers": {
-    "swapidev-ela": {
-      "auth": {}
-    }
-  }
-}
