--- fixtures/consumer/starwars/swapidev-ela.provider.json
@@ -1,10 +1,0 @@
-{
-	"name": "swapidev-ela",
-	"services": [
-		{
-			"baseUrl": "https://swapi.dev/api",
-			"id": "default"
-		}
-	],
-	"defaultService": "default"
-}
