--- fixtures/create/playground/superface/super.json
@@ -1,4 +1,0 @@
-{
-  "profiles": {},
-  "providers": {}
-}
